/go/vt/sqlparser/sql.go: fix parser

+++ go/vt/sqlparser/sql.go
@@ -1588,7 +1588,7 @@ const yyErrCode = 2
 const yyInitialStackSize = 16
 
 //line yacctab:1
-var yyExca = [...]int16{
+var yyExca = [...]int{
 	-1, 0,
 	1, 1238,
 	90, 1238,
@@ -2043,7 +2043,7 @@ const yyPrivate = 57344
 
 const yyLast = 82802
 
-var yyAct = [...]int16{
+var yyAct = [...]int{
 	758, 46, 4687, 4658, 4621, 714, 704, 1212, 4423, 4674,
 	576, 2968, 2486, 4274, 8, 4399, 4397, 4623, 4541, 4273,
 	7, 4527, 4271, 3, 4542, 4386, 1434, 4272, 6, 4305,
@@ -10327,7 +10327,7 @@ var yyAct = [...]int16{
 	0, 87,
 }
 
-var yyPact = [...]int32{
+var yyPact = [...]int{
 	82104, -1000, -501, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
@@ -10805,7 +10805,7 @@ var yyPact = [...]int32{
 	-1000, -1000, -1000,
 }
 
-var yyPgo = [...]int16{
+var yyPgo = [...]int{
 	0, 3586, 242, 1680, 22, 3584, 243, 328, 319, 26,
 	2880, 3583, 303, 3566, 298, 3564, 27, 19, 13, 31,
 	3559, 3190, 3188, 3185, 3183, 29, 3554, 8, 3550, 3548,
@@ -10856,7 +10856,7 @@ var yyPgo = [...]int16{
 	120, 3112, 3102, 216, 3100,
 }
 
-var yyR1 = [...]int16{
+var yyR1 = [...]int{
 	0, 465, 465, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
@@ -11138,7 +11138,7 @@ var yyR1 = [...]int16{
 	467,
 }
 
-var yyR2 = [...]int8{
+var yyR2 = [...]int{
 	0, 1, 2, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
@@ -11420,7 +11420,7 @@ var yyR2 = [...]int8{
 	1,
 }
 
-var yyChk = [...]int16{
+var yyChk = [...]int{
 	-1000, -465, -1, -4, -10, -15, -16, -17, -18, -19,
 	-21, -74, -22, -23, -24, -84, -88, -89, -90, -86,
 	-87, -94, -95, -96, -92, -91, -30, -33, -25, -98,
@@ -11898,7 +11898,7 @@ var yyChk = [...]int16{
 	83, 83, 207,
 }
 
-var yyDef = [...]int16{
+var yyDef = [...]int{
 	-2, -2, 1, 3, 4, 5, 6, 7, 8, 9,
 	10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
 	20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
@@ -12376,7 +12376,7 @@ var yyDef = [...]int16{
 	461, 464, 427,
 }
 
-var yyTok1 = [...]int16{
+var yyTok1 = [...]int{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -12392,7 +12392,7 @@ var yyTok1 = [...]int16{
 	3, 3, 3, 3, 129, 3, 142,
 }
 
-var yyTok2 = [...]int16{
+var yyTok2 = [...]int{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -12421,7 +12421,7 @@ var yyTok2 = [...]int16{
 	271, 272, 273, 274, 275, 276,
 }
 
-var yyTok3 = [...]uint16{
+var yyTok3 = [...]int{
 	57600, 277, 57601, 278, 57602, 279, 57603, 280, 57604, 281,
 	57605, 282, 57606, 283, 57607, 284, 57608, 285, 57609, 286,
 	57610, 287, 57611, 288, 57612, 289, 57613, 290, 57614, 291,
@@ -12600,9 +12600,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := int(yyPact[state])
+	base := yyPact[state]
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
+		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -12612,13 +12612,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
+		for yyExca[i] != -1 || yyExca[i+1] != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := int(yyExca[i])
+			tok := yyExca[i]
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -12649,30 +12649,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = int(yyTok1[0])
+		token = yyTok1[0]
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = int(yyTok1[char])
+		token = yyTok1[char]
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = int(yyTok2[char-yyPrivate])
+			token = yyTok2[char-yyPrivate]
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = int(yyTok3[i+0])
+		token = yyTok3[i+0]
 		if token == char {
-			token = int(yyTok3[i+1])
+			token = yyTok3[i+1]
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = int(yyTok2[1]) /* unknown char */
+		token = yyTok2[1] /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -12727,7 +12727,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = int(yyPact[yystate])
+	yyn = yyPact[yystate]
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -12738,8 +12738,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = int(yyAct[yyn])
-	if int(yyChk[yyn]) == yytoken { /* valid shift */
+	yyn = yyAct[yyn]
+	if yyChk[yyn] == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -12752,7 +12752,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = int(yyDef[yystate])
+	yyn = yyDef[yystate]
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -12761,18 +12761,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
+			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = int(yyExca[xi+0])
+			yyn = yyExca[xi+0]
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = int(yyExca[xi+1])
+		yyn = yyExca[xi+1]
 		if yyn < 0 {
 			goto ret0
 		}
@@ -12794,10 +12794,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
+				yyn = yyPact[yyS[yyp].yys] + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
-					if int(yyChk[yystate]) == yyErrCode {
+					yystate = yyAct[yyn] /* simulate a shift of "error" */
+					if yyChk[yystate] == yyErrCode {
 						goto yystack
 					}
 				}
@@ -12833,7 +12833,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= int(yyR2[yyn])
+	yyp -= yyR2[yyn]
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -12844,16 +12844,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = int(yyR1[yyn])
-	yyg := int(yyPgo[yyn])
+	yyn = yyR1[yyn]
+	yyg := yyPgo[yyn]
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = int(yyAct[yyg])
+		yystate = yyAct[yyg]
 	} else {
-		yystate = int(yyAct[yyj])
-		if int(yyChk[yystate]) != -yyn {
-			yystate = int(yyAct[yyg])
+		yystate = yyAct[yyj]
+		if yyChk[yystate] != -yyn {
+			yystate = yyAct[yyg]
 		}
 	}
 	// dummy call; replaced with literal code
